+++ Pal.Client/Floors/TerritoryState.cs
@@ -16,7 +16,7 @@ namespace Pal.Client.Floors
             _condition = condition;
         }
 
-        public ushort LastTerritory { get; set; }
+        public uint LastTerritory { get; set; }
         public PomanderState PomanderOfSight { get; set; } = PomanderState.Inactive;
         public PomanderState PomanderOfIntuition { get; set; } = PomanderState.Inactive;
